move root-layout to app folder
