+++ client/src/app/(root)/root-layout.tsx
